+++ packages/prop-house-contracts/contracts/starknet-cairo-v1/src/common/voting/vanilla.cairo
@@ -5,7 +5,7 @@ mod VanillaVotingStrategy {
 
     impl VanillaVotingStrategy of IVotingStrategy {
         fn get_voting_power(
-            timestamp: u32,
+            timestamp: u64,
             voter_address: ContractAddress,
             params: Array<felt252>,
             user_params: Array<felt252>,
@@ -16,7 +16,7 @@ mod VanillaVotingStrategy {
 
     #[external]
     fn get_voting_power(
-        timestamp: u32,
+        timestamp: u64,
         voter_address: ContractAddress,
         params: Array<felt252>,
         user_params: Array<felt252>,
